HiddenInput test to TS
