+++ src/signals/incident/components/form/HiddenInput/index.ts
@@ -1,0 +1,1 @@
+export { default } from './HiddenInput'
